Add docstring. Move tls options to http.
